add essay type question
The general idea here is to add a proper "essay" question type, without fiddling around with different "string" variants. I removed the exmaxchars part in read_metainfo as it throws an error and I do not undestand why maxchars should be transformed this way.

For now, essay only applies to qti21, the rest should work as normal.

This is experimental, so please use with caution.